# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "additionalProperties": true,
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemValidationError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 100.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "detail",
          "instance"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemValidationError": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": [
          "detail",
          "pointer"
        ],
        "title": "ProblemValidationError",
        "type": "object"
      },
...
      "RecruitmentStage": {
        "enum": [
          "daily_internship",
          "summer_internship",
          "winter_internship",
          "autumn_recruitment",
          "spring_recruitment",
          "other"
        ],
        "title": "RecruitmentStage",
        "type": "string"
      },
      "RequirementExplicitness": {
        "enum": [
          "explicit",
          "implicit",
          "uncertain"
        ],
        "title": "RequirementExplicitness",
        "type": "string"
      },
...
      "RequirementSummaryItemResponse": {
        "properties": {
          "confirmed_job_count": {
            "title": "Confirmed Job Count",
            "type": "integer"
          },
          "coverage_rate": {
            "title": "Coverage Rate",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/SummaryEvidenceResponse"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_count": {
            "title": "Evidence Count",
            "type": "integer"
          },
          "explicit_evidence_count": {
            "title": "Explicit Evidence Count",
            "type": "integer"
          },
          "mentioning_job_count": {
            "title": "Mentioning Job Count",
            "type": "integer"
          },
          "normalized_name": {
            "title": "Normalized Name",
            "type": "string"
          },
          "requirement_type": {
            "$ref": "#/components/schemas/RequirementType"
          }
        },
        "required": [
          "normalized_name",
          "requirement_type",
          "mentioning_job_count",
          "confirmed_job_count",
          "coverage_rate",
          "evidence_count",
          "explicit_evidence_count",
          "evidence"
        ],
        "title": "RequirementSummaryItemResponse",
        "type": "object"
      },
      "RequirementType": {
        "enum": [
          "eligibility",
          "core_competency",
          "experience",
          "preferred",
          "uncertain"
        ],
        "title": "RequirementType",
        "type": "string"
      },
      "SelectedSummaryRequest": {
        "properties": {
          "job_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 500,
            "minItems": 2,
            "title": "Job Ids",
            "type": "array"
          }
        },
        "required": [
          "job_ids"
        ],
        "title": "SelectedSummaryRequest",
        "type": "object"
      },
...
      "SummaryEvidenceResponse": {
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "explicitness": {
            "$ref": "#/components/schemas/RequirementExplicitness"
          },
          "job_id": {
            "format": "uuid",
            "title": "Job Id",
            "type": "string"
          },
          "job_title": {
            "title": "Job Title",
            "type": "string"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "requirement_id": {
            "format": "uuid",
            "title": "Requirement Id",
            "type": "string"
          }
        },
        "required": [
          "requirement_id",
          "job_id",
          "company_name",
          "job_title",
          "original_text",
          "explicitness"
        ],
        "title": "SummaryEvidenceResponse",
        "type": "object"
      },
      "TargetRoleCreate": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "recruitment_stage": {
            "$ref": "#/components/schemas/RecruitmentStage"
          }
        },
        "required": [
          "name",
          "recruitment_stage"
        ],
        "title": "TargetRoleCreate",
        "type": "object"
      },
      "TargetRoleListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TargetRoleResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "TargetRoleListResponse",
        "type": "object"
      },
      "TargetRoleResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "job_count": {
            "default": 0,
            "title": "Job Count",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "recruitment_stage": {
            "$ref": "#/components/schemas/RecruitmentStage"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "recruitment_stage",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "TargetRoleResponse",
        "type": "object"
      },
      "TargetRoleSummaryResponse": {
        "properties": {
          "confirmed_job_count": {
            "title": "Confirmed Job Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/RequirementSummaryItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "sample_job_count": {
            "title": "Sample Job Count",
            "type": "integer"
          },
          "sample_size_notice": {
            "title": "Sample Size Notice",
            "type": "string"
          },
          "selected_job_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Selected Job Ids",
            "type": "array"
          },
          "target_role_id": {
            "format": "uuid",
            "title": "Target Role Id",
            "type": "string"
          },
          "target_role_name": {
            "title": "Target Role Name",
            "type": "string"
          }
        },
        "required": [
          "target_role_id",
          "target_role_name",
          "sample_job_count",
          "confirmed_job_count",
          "sample_size_notice",
          "items"
        ],
        "title": "TargetRoleSummaryResponse",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "岗位门槛分析系统 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/target-roles": {
      "get": {
        "operationId": "list_target_roles_endpoint_api_v1_target_roles_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRoleListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "List Target Roles Endpoint",
        "tags": [
          "target-roles"
        ]
      },
      "post": {
        "operationId": "create_target_role_endpoint_api_v1_target_roles_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetRoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Create Target Role Endpoint",
        "tags": [
          "target-roles"
        ]
      }
    },
    "/api/v1/target-roles/{role_id}": {
      "delete": {
        "operationId": "delete_target_role_endpoint_api_v1_target_roles__role_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Delete Target Role Endpoint",
        "tags": [
          "target-roles"
        ]
      },
      "get": {
        "operationId": "get_target_role_endpoint_api_v1_target_roles__role_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRoleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Get Target Role Endpoint",
        "tags": [
          "target-roles"
        ]
      }
    },
...
    "/api/v1/target-roles/{role_id}/summary": {
      "get": {
        "operationId": "get_target_role_summary_endpoint_api_v1_target_roles__role_id__summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requirement_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RequirementType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requirement Type"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRoleSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Get Target Role Summary Endpoint",
        "tags": [
          "summary"
        ]
      },
      "post": {
        "operationId": "create_selected_target_role_summary_endpoint_api_v1_target_roles__role_id__summary_post",
        "parameters": [
          {
            "in": "path",
            "name": "role_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Role Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "requirement_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RequirementType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requirement Type"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectedSummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetRoleSummaryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Create Selected Target Role Summary Endpoint",
        "tags": [
          "summary"
        ]
      }
    },
...
  }
}